requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.10.0.84
ultralytics==8.3.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.1
requests>=2.31.0
python-dotenv>=1.0.0
aiofiles==23.2.1
httpx>=0.28.0
numpy>=1.23.0,<2.0.0
pydantic>=2.0.0,<3.0.0
